+++ Source/UniqueEngine/Graphics/GraphicsBuffer.h
@@ -15,7 +15,7 @@ namespace Unique
 		template<class T>
 		bool Create(const Vector<T>& data, Usage usage = Usage::USAGE_STATIC)
 		{
-			return Create(data.size(), sizeof(T), usage, (void*)data.data());
+			return Create((uint)data.size(), sizeof(T), usage, (void*)data.data());
 		}
 
 		template<class T>
